Rename EventInterpreter to BoardSyncer, and fully encapsulate syncing logic within it.
