+++ HouseRules_Core/HouseRules_Core.csproj
@@ -47,7 +47,7 @@
         </Reference>
     </ItemGroup>
     <ItemGroup>
-        <Compile Include="EventInterpreter.cs" />
+        <Compile Include="BoardSyncer.cs" />
         <Compile Include="ModPatcher.cs" />
         <Compile Include="Properties\AssemblyInfo.cs" />
         <Compile Include="HR.cs" />
